app/requirements.txt
numpy
pandas
seaborn
matplotlib
opencv-python
Pillow